change the JSX Open/Closing Tag Background Color
When you click on the open tag in JSX you should see a highlighted background of the closing and opening Tags but It is barely visible. You could only see a good contrast background if you double-click on the tag name, so I changed the color for the single press (editor.wordHighlightBackground) to be more visible.